Changed metadata size to allow for longer messages

+++ StegApp/headers/Image.h
@@ -2,7 +2,7 @@
 #include<QString>
 //defines the size of the header, has to be a set size for decoding purposes
 //header is where metadata is stored
-#define STEG_HEADER_SIZE sizeof(uint32_t) * 8
+#define STEG_HEADER_SIZE sizeof(uint64_t) * 8
 #define MESSAGE_LEN_HEADER STEG_HEADER_SIZE / 2
 #define NOISE_AMOUNT_HEADER STEG_HEADER_SIZE / 2
